hw3/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
pycocotools>=2.0.6
tensorboard>=2.14.0
torch-tb-profiler>=0.4.0
diffusers>=0.21.0
accelerate>=0.23.0
controlnet-aux>=0.0.6
opencv-python>=4.8.0
seaborn>=0.12.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
tqdm>=4.65.0
gradio>=3.50.0
safetensors>=0.3.0
timm>=0.9.0
scikit-learn==1.7.2
matplotlib==3.5.1
albumentations==2.0.8
jinja2